services/content-service/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "content-service"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "sqlalchemy[asyncio]>=2.0",
    "psycopg[binary]>=3.1",
    "pydantic-settings>=2.0",
    "pdfplumber>=0.11",
    "python-docx>=1.1",
    "tiktoken>=0.7",
    "sentence-transformers>=2.2",
    "common",
]

[project.optional-dependencies]
dev = ["ruff>=0.4", "black>=24", "pytest>=8", "pytest-asyncio>=0.23", "httpx>=0.27"]

[tool.ruff]
target-version = "py312"
line-length = 100
select = ["E", "F", "I", "UP"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
